std::ostream and gtest print tests, invScaled bug fix.
Add gtest printing tests and support gtest printing via adapting
llvm::raw_ostream to std::ostream.
Fixed a bug in scaledInv, which would fail if the diagonal matrix's
diagonal elements were all `1` or `-1`, and at least one of them
was `-1`. When the lcm was `1`, it'd skip scaling the result matrix,
but that missed the fact that some rows needed their signs flipped.