Use [inner-outer] representation consistently
In particular, Phi was using [outer<->inner] for rows (representing
final schedule order), but [inner<->outer] for cols (representing the
original indvars being mapped). On top of the confusion resulting from
inconsistency, it meant that the exchange matrix was actually the
identity w/ respect to affine transforms. Which meant that Phi needed
to be multiplied with an exchange matrix itself to actually preserve
mappings, in which case it was easier to just fix our representaiton.

Also, make `end` consistent by introducing `last = end-1`.