WIP Costmodeling (#80)
* make members private

* minor progress

* house cleaning

* remove bad code

* negligible dev

* minor progress; tests currently fail because x.data() is not constexpr

* tests pass

* subscript vectors

* minor progress

* added a few comments

* add inv function

* format

* fix vgcd, remove comment, and add check to skip computation in solveSystem

* Add inv and scaledInv, test them.

* std::ostream and gtest print tests, invScaled bug fix.

Add gtest printing tests and support gtest printing via adapting
llvm::raw_ostream to std::ostream.
Fixed a bug in scaledInv, which would fail if the diagonal matrix's
diagonal elements were all `1` or `-1`, and at least one of them
was `-1`. When the lcm was `1`, it'd skip scaling the result matrix,
but that missed the fact that some rows needed their signs flipped.

* Avoid false circular dependency for std::ostream adaptation.

* pch

* update ci

* flatten broadcasts and indexing, non-nullptr, and adjust gtest link

* use some std::algs instead of loops

* BitSets for referencing Dependences in MemoryAccesses (#81)

* Have MemoryAccess hold a vector of pointers to dependencies

(as opposed to a vector of indices into a vector of dependencies)

* Seems using integers MemoryAccesses to refer to edges is best

This is because of heavy use of `BitSets` in the LP code for marking
active edges; all these set operations are cheap with `BitSets`, but
would be relatively expensive using `SmallPtrSet`

* tests pass

* format cmake

* Remove problematic header.

* Apple doesn't have std::ranges

* use rtti in tests when using ubsan

* ` -DCMAKE_BUILD_TYPE=Debug` 

Forgot ` -DCMAKE_BUILD_TYPE=Debug` when updating CI earliear

* MemoryAccess private

* Add new file

* Inlinememoryaccess (#83)

* Start having MemoryAccess be defined contiguously.

* Finish making MemoryAccess a contiguous block in memory.

* FAM base length 1

* Add benchmarks again.

* Address new Wshadow warnings

* Start updating to use ScheduledMemoryAccess

`Instruction`s will hold these instead of `MemoryAccess`es.
The `ScheduledMemoryAccess` should hold relevant meta-data to
aid in cost modeling and eventual code generation.

* Add libpfm for google bench

* Fix benchmarks

* Add identity lower bound fun, and install tex for doxy

* Tests pass again locally

* allign allocations, renamed ScheduledMemoryAccess to Address

* Experiment with AddressSanitizer. LLVM appears to leak a lot

But I did find a couple problems within the library:
1. `alloc.Allocate<char>(nbytes)` is going to 1-byte align data;
   I need to pass the actual alignment needed to two-arg
   `alloc.Allocate(nbytes, alignment)`.
2. `alloc.Allocate<T>` does not call a constructor or initialize
   memory in any way, thus the fields of `T` may be in a totally
   invalid state, e.g. a vector with invalid pointer and capacity
   entries, which means it can't be recovered via calling `clear()`.
   Need to explicitly call the construct into this memory, e.g.
   `pt = alloc.Allocate<T>(); ret = new(pt) T(args...)`.

* Add `Testing` to gitignore, and add leak suppressions file

Need to look more at these leaks to see to what extant they
originate from or maybe can be addressed from this library.

* Use [inner-outer] representation consistently

In particular, Phi was using [outer<->inner] for rows (representing
final schedule order), but [inner<->outer] for cols (representing the
original indvars being mapped). On top of the confusion resulting from
inconsistency, it meant that the exchange matrix was actually the
identity w/ respect to affine transforms. Which meant that Phi needed
to be multiplied with an exchange matrix itself to actually preserve
mappings, in which case it was easier to just fix our representaiton.

Also, make `end` consistent by introducing `last = end-1`.

* Make configure.benchmark executable.

* MemoryAccess now is_trivially_copyable_v

* Add `static_assert`

* Refactor, splitting Math.hpp into a few pieces

As the project grows, I think it's useful to be a bit more organized.
Jumping to locations may be a little easier across vs within files.

* `#ifndef NDEBUG` was at wrong location