Minor tweak to iteration
Currently still need to debug removeLoop working seemingly
incorrectly, and then memory issues by re-enabling the sanitizer.
There are a few issues probably resulting from under allocation in
comparator, but also LLVM doesn't seem to be cleaning up data we
allocate in tests. This may be because we're making a mistake in how
we build up the llvm data structures, so that the usual destructors
that get called somehow end up missing the pointers we made.