Some updates, have a general plan for generating new IR
We postpone truncation towards the end, each layer of the recursion
returns a loop and number excluded. If the number excluded exceeds
current depth, then this level has a problem and is not feasible,
e.g. one of these loops returning from the recursion has an array
indexed in a non-affine way, or that loop itself doesn't have simple
bounds.
This inside-out generation approach leads to a much simpler
conditioning/parsing stream than current.

Whenever we get sets of loops, we update our IR to exclude all the
excluded, and run the analysis before continuing on. This lets us free
the associated IR memory and reuse for the next cluster.