Addr updates to support in place `rotate`
Store one `Addr` field in `TreeResult`, add iterate methods to
separate loads and stows.
Track maxdepth the TreeResult encountered, and use this to have all
`Addr`s allocate enough space to grow on `rotate`.
Replace the construction of new `Addr`s with an in-place rotation,
because we wouldn't need the old one anymore, and to keep all the
edges in tact.