Data-oriented memory layout (#96)
* Start adding IR file

* Add some more comments

* Minor progress

* negligible changes

* Totally broken state; figuring out basic IR structure

* Minor progress; prepping to combine ArrayIndex and Addr

* Remove type parameter from AffineLoopNest

It is now a runtime value whether the lower bounds are 0; we should be
able to save on allocations/copies.

* a fix

* some progress on restructuring

* minor additions; will probably replace LoopTree, too

* minor additions

* minor additions

* negligible additions

* Add UList and remove llvm dependency from BumpAlloc

* Some more minor cleanup, moving SmallSparseMatrix to own file

* More constraints to ostream concept

* drop more llvms

* Some updates, have a general plan for generating new IR

We postpone truncation towards the end, each layer of the recursion
returns a loop and number excluded. If the number excluded exceeds
current depth, then this level has a problem and is not feasible,
e.g. one of these loops returning from the recursion has an array
indexed in a non-affine way, or that loop itself doesn't have simple
bounds.
This inside-out generation approach leads to a much simpler
conditioning/parsing stream than current.

Whenever we get sets of loops, we update our IR to exclude all the
excluded, and run the analysis before continuing on. This lets us free
the associated IR memory and reuse for the next cluster.

* Approach needs to change...

* Some progress

* Starting on instr graph

* CMake tweaks

* Separate math library (#94)

* No math

* Some include updates

* add Stow and Load subtypes

* pretty negligible progress, actually

* minor cleanup

* minor...

* minor progress

* Further refactoring progress

* some progress on updating Predicate::Set

* Minor progress on predicates

* some updates

* minor cleanup

* some progress

* minor progress

* More minor progress

* updating instructions

* minor progress

* some more progress

* Minor progress on replaceAllUsesWith implementation

* some progress on cse

* further updates

* reduce unnecessary map loopups

* some more minor progress

* more minor progress

* some further progress

* a few more updates on Instruction

* Added Value type to distinguish from Loop and Exit

* more minor progress

* minor changes

need to fix setVal/setOperands and updating users story
look at replaceUsesByUsers and a setOperands function

* bleh

* Updated ControlFlowMerging

* next change is that loop invariants need to be removed from cache

and handled in the TreeResults instead

* decent progress

* minor changes

* minor progress

The immediate plan here is to add an iterator for ancestors in control flow
merging, and to update the BBpath code to set TreeResults.
Additionally, will probably sort Addrs to be [Stows..., Loads...]

* Some progress on setting up parsing correctly

* no need for bbpred map vector, can just use map

* A few updates

Use `MapVector` for BBPredMap afterall

* minor progress

* Some progress updating for changes, and added addPrediate

* some more minor progress

* Update graphs

* TurboLoop simplifications

* no exit

* Minor progress, delete a couple files

* starting to work on LP LoopBlock

* Arena updates

* Some minor progress updating dep graph building

* add a comment

* Simplify `users` and optimize for common case of 1 user

* Some progress updating LoopBlock for the new IR

* Some further progress updating linear program

* negligible

* minor changes

* minor progress

* some progress updating LoopBlock

* Some more comments

* Need to redefine graphs interface

* update graphs

* minor progress

* minor updates

* minor tweaks; need to figure out Graph plan

Probably take more advantage of `std::ranges` for creating nested graph
iterators, in particular `std::views::filter` looks useful, e.g. for
iterating only over edges unsatisfied at a given level.

* NestedList updates

* Add filtered depth iterators

* graph updates

* update `breakGraph`

* resolve warnings in LoopBlock, but file still a WIP

* Return result enum from lp loop block

* check if stashed prevents reordering before trying to sat that dep

* Minor progress

* minor progress

* Addr updates to support in place `rotate`

Store one `Addr` field in `TreeResult`, add iterate methods to
separate loads and stows.
Track maxdepth the TreeResult encountered, and use this to have all
`Addr`s allocate enough space to grow on `rotate`.
Replace the construction of new `Addr`s with an in-place rotation,
because we wouldn't need the old one anymore, and to keep all the
edges in tact.

* update LoopBlock

* minor progress

* some comments on top sort alg

* minor progress on top sort

* more topsort progress

* Minor progress

* starting addr simplification

* Minor progress on updating Address

* A few DependencePolyhedra fixes

* some more minor progress, add a few more comments on planned approach

* Add more comments

* Some further progress; short term, lets get old tests passing

Before working on cost modeling and load elimination.

* some more progress on updating tests

* minor progress

* negligible

* minor changes

* start adding `Dependencies` struct

* start trying dod

* more minor progress

* minor progress

* undelete

* Fix diagnostics in Dependence.hpp

* minor progress

* A few more updates

* update sat stash

* finish updating stash sat